Implanta Depósito, funções de investimento e abre conta poupança

+++ src/br/gov/caixa/app/Models/Services/Saque/ConsultarSaldo.java
@@ -3,7 +3,8 @@ package br.gov.caixa.app.Models.Services.Saque;
 import br.gov.caixa.app.Models.Common.Conta;
 
 public class ConsultarSaldo {
-    public static Double consultaSaldo(Conta conta){
-        return (double) conta.getSaldo();
+    public static float consultaSaldo(Conta conta){
+
+        return conta.getSaldo();
     }
 }
